Bring routes and app specific components back to apps
