+++ apps/catalog/catalog/src/app/marketplace/title/view/view.module.ts
@@ -5,12 +5,34 @@ import { CommonModule } from '@angular/common';
 // Component
 import { MarketplaceMovieViewComponent } from './view.component';
 import { MovieShellModule } from '@blockframes/movie/marketplace/shell/shell.module';
+import { RouterModule } from '@angular/router';
+
+const routes = [{
+  path: '',
+  component: MarketplaceMovieViewComponent,
+  children: [
+    {
+      path: '',
+      redirectTo: 'main',
+      pathMatch: 'full'
+    },
+    {
+      path: 'main',
+      loadChildren: () => import('@blockframes/movie/marketplace/main/main.module').then(m => m.MovieMainModule)
+    },
+    {
+      path: 'avails',
+      loadChildren: () => import('../avails/avails.module').then(m => m.MarketplaceMovieAvailsModule)
+    }
+  ]
+}];
 
 @NgModule({
   declarations: [MarketplaceMovieViewComponent],
   imports: [
     CommonModule,
     MovieShellModule,
+    RouterModule.forChild(routes)
   ]
 })
 export class MovieViewModule {}
